+++ tests/test.R
@@ -1,5 +1,5 @@
 if(FALSE){
-  library(GLmomentEst)
+  library(GLmom)
   data("streamflow")
   glme.gev(streamflow$r1)
   ?glme.gev